+++ node-net-manager/network/Net_test.go
@@ -1,6 +1,7 @@
 package network
 
 import (
+	"net"
 	"testing"
 
 	"gotest.tools/assert"
@@ -10,7 +11,7 @@ func TestPortMappingEmpty(t *testing.T) {
 	mock := &mockiptable{}
 	iptable = mock
 	//empty string
-	err := ManageContainerPorts("0.0.0.0", "", OpenPorts)
+	err := ManageContainerPorts(net.ParseIP("0.0.0.0"), "", OpenPorts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -21,7 +22,7 @@ func TestPortMappingUdp(t *testing.T) {
 	mock := &mockiptable{}
 	iptable = mock
 	//udp 80
-	err := ManageContainerPorts("0.0.0.0", "80:80/udp", OpenPorts)
+	err := ManageContainerPorts(net.ParseIP("0.0.0.0"), "80:80/udp", OpenPorts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -29,7 +30,7 @@ func TestPortMappingUdp(t *testing.T) {
 		assert.Equal(t, arg, mock.CalledWith[i])
 	}
 	//udp 80 and 90
-	err = ManageContainerPorts("0.0.0.0", "80:80/udp;90:100/udp", OpenPorts)
+	err = ManageContainerPorts(net.ParseIP("0.0.0.0"), "80:80/udp;90:100/udp", OpenPorts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -42,7 +43,7 @@ func TestPortMappingTcp(t *testing.T) {
 	mock := &mockiptable{}
 	iptable = mock
 	//tcp 80
-	err := ManageContainerPorts("0.0.0.0", "80", OpenPorts)
+	err := ManageContainerPorts(net.ParseIP("0.0.0.0"), "80", OpenPorts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -50,7 +51,7 @@ func TestPortMappingTcp(t *testing.T) {
 		assert.Equal(t, arg, mock.CalledWith[i])
 	}
 	//tcp 80:80
-	err = ManageContainerPorts("0.0.0.0", "80:80", OpenPorts)
+	err = ManageContainerPorts(net.ParseIP("0.0.0.0"), "80:80", OpenPorts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -58,7 +59,7 @@ func TestPortMappingTcp(t *testing.T) {
 		assert.Equal(t, arg, mock.CalledWith[i])
 	}
 	//tcp 80 and 90
-	err = ManageContainerPorts("0.0.0.0", "80:80/tcp;90:100/tcp", OpenPorts)
+	err = ManageContainerPorts(net.ParseIP("0.0.0.0"), "80:80/tcp;90:100/tcp", OpenPorts)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -71,17 +72,17 @@ func TestPortMappingInvalid(t *testing.T) {
 	mock := &mockiptable{}
 	iptable = mock
 
-	err := ManageContainerPorts("0.0.0.0", "80:80-80", OpenPorts)
+	err := ManageContainerPorts(net.ParseIP("0.0.0.0"), "80:80-80", OpenPorts)
 	if err == nil {
 		t.Fatal("80:80-80 must be invalid")
 	}
 
-	err = ManageContainerPorts("0.0.0.0", " ", OpenPorts)
+	err = ManageContainerPorts(net.ParseIP("0.0.0.0"), " ", OpenPorts)
 	if err == nil {
 		t.Fatal("space must be invalid")
 	}
 
-	err = ManageContainerPorts("0.0.0.0", "hello", OpenPorts)
+	err = ManageContainerPorts(net.ParseIP("0.0.0.0"), "hello", OpenPorts)
 	if err == nil {
 		t.Fatal("hello must be invalid")
 	}
